+++ sdk/core/azure-core/test/ut/http.cpp
@@ -161,8 +161,8 @@ namespace Azure { namespace Core { namespace Test {
       Http::Url url("http://test.com");
       Http::Request req(httpMethod, url);
 
-      Azure::IO::_internal::NullBodyStream* d
-          = dynamic_cast<Azure::IO::_internal::NullBodyStream*>(req.GetBodyStream());
+      Azure::Core::IO::_internal::NullBodyStream* d
+          = dynamic_cast<Azure::Core::IO::_internal::NullBodyStream*>(req.GetBodyStream());
       EXPECT_TRUE(d);
 
       req.StartTry();
@@ -178,7 +178,7 @@ namespace Azure { namespace Core { namespace Test {
 
       EXPECT_FALSE(headers.count("name"));
 
-      d = dynamic_cast<Azure::IO::_internal::NullBodyStream*>(req.GetBodyStream());
+      d = dynamic_cast<Azure::Core::IO::_internal::NullBodyStream*>(req.GetBodyStream());
       EXPECT_TRUE(d);
     }
 
@@ -187,31 +187,33 @@ namespace Azure { namespace Core { namespace Test {
       Http::Url url("http://test.com");
 
       std::vector<uint8_t> data = {1, 2, 3, 4};
-      Azure::IO::MemoryBodyStream stream(data);
+      Azure::Core::IO::MemoryBodyStream stream(data);
 
       // Change the offset of the stream to be non-zero by reading a byte.
       std::vector<uint8_t> temp(2);
       EXPECT_EQ(
-          Azure::IO::BodyStream::ReadToCount(stream, temp.data(), 1, GetApplicationContext()), 1);
+          Azure::Core::IO::BodyStream::ReadToCount(stream, temp.data(), 1, GetApplicationContext()),
+          1);
 
       EXPECT_EQ(temp[0], 1);
       EXPECT_EQ(temp[1], 0);
 
       Http::Request req(httpMethod, url, &stream);
 
-      Azure::IO::MemoryBodyStream* d
-          = dynamic_cast<Azure::IO::MemoryBodyStream*>(req.GetBodyStream());
+      Azure::Core::IO::MemoryBodyStream* d
+          = dynamic_cast<Azure::Core::IO::MemoryBodyStream*>(req.GetBodyStream());
       EXPECT_TRUE(d);
 
       req.StartTry();
 
-      d = dynamic_cast<Azure::IO::MemoryBodyStream*>(req.GetBodyStream());
+      d = dynamic_cast<Azure::Core::IO::MemoryBodyStream*>(req.GetBodyStream());
       EXPECT_TRUE(d);
 
       // Verify that StartTry rewound the stream back.
       auto getStream = req.GetBodyStream();
       EXPECT_EQ(
-          Azure::IO::BodyStream::ReadToCount(*getStream, temp.data(), 2, GetApplicationContext()),
+          Azure::Core::IO::BodyStream::ReadToCount(
+              *getStream, temp.data(), 2, GetApplicationContext()),
           2);
 
       EXPECT_EQ(temp[0], 1);