+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -48,7 +48,7 @@ TEST(SimplifiedHeader, core)
 
   {
     std::vector<uint8_t> buffer(10);
-    EXPECT_NO_THROW(Azure::IO::MemoryBodyStream mb(buffer));
+    EXPECT_NO_THROW(Azure::Core::IO::MemoryBodyStream mb(buffer));
   }
   EXPECT_NO_THROW(Azure::Core::Http::TelemetryPolicy tp("", ""));
 }